requirements.txt
aiofiles==23.2.1
django-extensions==3.2.3
drf-nested-routers==0.94.1
gradio==5.16.1
gradio-client==1.7.0
markupsafe==2.1.5
pydotplus==2.0.2
ruff==0.9.6
safehttpx==0.1.6
tomlkit==0.13.2
websockets==14.2
requests==2.32
djangorestframework==3.15.2
django==5.1.6
django-debug-toolbar==5.0.1
drf-spectacular==0.28.0
inflection==0.5.1
django-countries==7.6.1
django-import-export==4.3.6 
tablib==3.8.0
djangorestframework-datatables==0.7.2
django-grappelli
django-guardian
